Make reusable component for dashboards
